fix: file clean up. Api modules separated. Profile edit form fix. Add new rollercoaster form fix url encoding query params.
