+++ src/__tests__/profile/ProfileList.test.js
@@ -4,12 +4,12 @@ import ProfileList from "../../components/profile/ProfileList";
 import { Router } from "react-router-dom";
 import { useAuth0 } from "../../contexts/react-auth0-context";
 import { createMemoryHistory } from "history";
-// import ApiManager from "../../modules/ApiManager";
+import userManager from "../../modules/users/userManager";
 // import { fireEvent, act } from "@testing-library/react";
 
 const history = createMemoryHistory();
 jest.mock("../../contexts/react-auth0-context");
-// jest.mock("../../modules/ApiManager");
+jest.mock("../../users/userManager");
 /* throws uncaught exception if not mocked properly */
 
 describe("<ProfileList />", () => {
@@ -22,8 +22,6 @@ describe("<ProfileList />", () => {
       picUrl: "picture goes here"
     },
     history: []
-    // editProfile: jest.fn()
-    // addNewCredit: jest.fn()
   };
 
   const user = {
@@ -40,39 +38,39 @@ describe("<ProfileList />", () => {
     });
   });
 
-  //   ApiManager.getUserProfile.mockReturnValue({
-  //       id: 1,
-  //       first_name: "matt",
-  //       last_name: "crook",
-  //       credits: [{ rollerCoasterId: 1}, {rollerCoasterId: 2}],
-  //       picUrl: "picture goes here"
-  //   })
+    userManager.getUserProfile.mockReturnValue({
+        id: 1,
+        first_name: "matt",
+        last_name: "crook",
+        credits: [{ rollerCoasterId: 1}, {rollerCoasterId: 2}],
+        picUrl: "picture goes here"
+    })
 
   it("Renders edit, delete profile buttons/ roller coaster cards", async () => {
     // mock out window.fetch for the test
-    // const fakeUserProfile = {
-    //   id: 1,
-    //   first_name: "matt",
-    //   last_name: "crook",
-    //   credits: [{ rollerCoasterId: 1 }, { rollerCoasterId: 2 }],
-    //   picUrl: "picture goes here"
-    // };
-    // jest.spyOn(ApiManager, "getUserProfile").mockImplementationOnce(() => {
-    //   return Promise.resolve({
-    //     json: () => Promise.resolve(fakeUserProfile)
-    //   });
-    // });
+    const fakeUserProfile = {
+      id: 1,
+      first_name: "matt",
+      last_name: "crook",
+      credits: [{ rollerCoasterId: 1 }, { rollerCoasterId: 2 }],
+      picUrl: "picture goes here"
+    };
+    jest.spyOn(userManager, "getUserProfile").mockImplementationOnce(() => {
+      return Promise.resolve({
+        json: () => Promise.resolve(fakeUserProfile)
+      });
+    });
 
 
     const profileListComponent = render(
-        <Router history={history}>
+      <Router history={history}>
         <ProfileList {...props} />
       </Router>
     );
 
 
-    // let creditsArray = fakeUserProfile.credits;
-    // expect(fakeUserProfile.credits).toEqual(creditsArray);
+    let creditsArray = fakeUserProfile.credits;
+    expect(fakeUserProfile.credits).toEqual(creditsArray);
 
 
     // console.log(profileListComponent.debug());
@@ -80,28 +78,12 @@ describe("<ProfileList />", () => {
     expect(profileListComponent).toBeTruthy();
 
     const profileCardTag = profileListComponent.getByTestId("profile_card_container_testid");
-    /* const deleteBtnTag = profileListComponent.getByTestId("delete_profile_btn_testid"); */
+    const deleteBtnTag = profileListComponent.getByTestId("delete_profile_btn_testid");
 
     expect(profileCardTag).toBeInTheDocument();
     expect(deleteBtnTag).toBeInTheDocument();
 
-    // const editBtnTag = profileListComponent.getByTestId("edit_profile_btn_testid");
-    // expect(editBtnTag).toBeInTheDocument();
-
-    // act(() => {
-        /* fire events that update state */
-        // fireEvent.click(profileListComponent.getByTestId("delete_profile_btn_testid"));
-      });
-      /* assert on the output */
-    // expect(ApiManager.deleteUserProfile).toHaveBeenCalledTimes(1);
-
-    // fireEvent.click(profileListComponent.getByTestId("edit_profile_btn_testid"));
-    // expect(props.editProfile).toHaveBeenCalled();
-
-    // fireEvent.click(profileListComponent.getByTestId("add_new_credit_btn_testid"));
-    // fireEvent.click(profileListComponent.getByTestId("edit_profile_btn_testid"));
-
-    // expect(props.addNewCredit).toHaveBeenCalledTimes(1);
-    // expect(props.editProfile).toHaveBeenCalledTimes(1);
-//   });
+    const editBtnTag = profileListComponent.getByTestId("edit_profile_btn_testid");
+    expect(editBtnTag).toBeInTheDocument();
+  });
 });
