+++ src/components/ApplicationViews.js
@@ -6,8 +6,7 @@ import LandingPage from "./auth/Login";
 import Home from "./home/Home";
 import NewRollerCoaster from "./addNewForm/NewRollerCoaster";
 import AddNewCreditForm from "./profile/NewCreditForm";
-import CreateAccount from "./auth/Register";
-import EditCreditForm from "./profile/EditCreditForm";
+// import CreateAccount from "./auth/Register";
 import MessageList from "./messages/Messages";
 import EditProfile from "./profile/EditProfile";
 import LeaderBoard from "./leaderBoard/LeaderBoard";
@@ -121,22 +120,6 @@ const ApplicationViews = ({
         render={(props) => {
           if (isAuthenticated === true && userProfile.id) {
             return <NewRollerCoaster {...props} />;
-          } else {
-            return <LandingPage />;
-          }
-        }}
-      />
-
-      <Route
-        path="/users/:creditId(\d+)/edit"
-        render={(props) => {
-          if (isAuthenticated === true && userProfile.id) {
-            return (
-              <EditCreditForm
-                rollerCoasterId={parseInt(props.match.params.rollerCoasterId)}
-                {...props}
-              />
-            );
           } else {
             return <LandingPage />;
           }
