+++ src/components/leaderBoard/LeaderBoard.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import ApiManager from "../../modules/ApiManager";
+import userManager from "../../modules/users/userManager";
 import { useAuth0 } from "../../contexts/react-auth0-context";
 import "./LeaderBoard.css";
 
@@ -10,9 +10,12 @@ const LeaderBoard = (props) => {
   const defaultQPicture = "https://cdn.dribbble.com/users/2908839/screenshots/6292457/shot-cropped-1554473682961.png";
   const defaultProfilePicture = "https://aesusdesign.com/wp-content/uploads/2019/06/mans-blank-profile-768x768.png";
 
+
+  // Function to get all users, build a custom object with the credits of each user attached to it, and set that as the profile in state.
+  // Then sort the profiles by the length of credits array from most to least.
   const getAllUsers = async () => {
     try {
-      const users = await ApiManager.getAllUsers();
+      const users = await userManager.getAllUsers();
       const profileWithCreditsArray = users.map((user) => {
         const profile = user.user;
         return {
@@ -37,7 +40,6 @@ const LeaderBoard = (props) => {
     getAllUsers();
   }, []);
 
-  console.log(userProfile)
 
   return (
     <>
