+++ src/components/profile/Profile.css
@@ -24,7 +24,7 @@
   transition: 0.3s;
   border-radius: 5px;
   background: rgb(233, 232, 232);
-  width: 23%;
+  width: 24%;
   margin-top: 10px;
   margin-bottom: 12px;
   margin-left: 10px;
