+++ src/components/search/CustomizedInputBase.js
@@ -1,10 +1,8 @@
-import React, { useState, useEffect } from "react";
-import ApiManager from "../../modules/ApiManager";
-import { useAuth0 } from "../../contexts/react-auth0-context";
+import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import MenuIcon from "@material-ui/icons/Menu";
 import SearchIcon from "@material-ui/icons/Search";
-import DirectionsIcon from "@material-ui/icons/Directions";
+// import DirectionsIcon from "@material-ui/icons/Directions";
 import Paper from "@material-ui/core/Paper";
 import InputBase from "@material-ui/core/InputBase";
 import Divider from "@material-ui/core/Divider";
