+++ src/modules/parks/parkManager.js
@@ -1,0 +1,42 @@
+const remoteURL = "http://localhost:8000";
+// const remoteURL = process.env.REACT_APP_BASE_URL;
+
+const parkManager = {
+  async postNewPark(resource) {
+    const data = await fetch(`${remoteURL}/parks`, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: "JWT" + localStorage.getItem("accessToken"),
+      },
+      body: JSON.stringify(resource),
+    });
+    return await data.json();
+  },
+
+  async getParks() {
+    const resp = await fetch(`${remoteURL}/parks`, {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/x-www-form-urlencoded",
+        Authorization: "JWT" + localStorage.getItem("accessToken"),
+      },
+      Accept: "application/json",
+    });
+    return await resp.json();
+  },
+
+  async getParkByName(park) {
+    const resp = await fetch(`${remoteURL}/parks?name=${park}`, {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/x-www-form-urlencoded",
+        Authorization: "JWT" + localStorage.getItem("accessToken"),
+      },
+      Accept: "application/json",
+    });
+    return await resp.json();
+  },
+};
+
+export default parkManager;
